Support instance creation w/out specifying type
The `createInstance` functions now take the JSON as the _first_ argument, and the type as an optional second argument.  The code will try to determine the type from the JSON, but if it can't, it will fall back on the type passed in (if relevant).

Also bump version to 5.2.0-alpha.2.