CIMPL6 support: concept and implied references
In CIMPL6, we do away with code/Coding/CodeableConcept and always use new primitive concept.  It's physical representation, however, is similar to CodeableConcept.

In CIMPL4 we also get rid of RefValue.  Instead, any field that points to an EntryElement is considered to be a reference.

Last, value is now always 1..1, so test cases around 0..1 and 0..* values have been removed.