Switch to built-in EntryInfo class
With removal of shr.base.Entry, the core entry metadata (SHR ID, Entry ID, Entry Type) are moved to a built-in EntryInfo class.

In addition, entry info is simplified so that it's properties have normal string values rather than objects (`{ Value: "my value" }`).  The entry information in references was also simplified to be string values.